Create a proper JavaStringToRubyStringNode and use it in ForeignToRubyNode
